Refactoring project shared packages
