+++ androidApp/src/main/java/com/leandrodev/contasdolar/android/auth/signin/AuthScreen.kt
@@ -18,9 +18,9 @@ import com.leandrodev.contasdolar.android.GoogleSignInButton
 import com.leandrodev.contasdolar.android.R
 import com.leandrodev.contasdolar.android.auth.google.GoogleSignInLauncher
 import com.leandrodev.contasdolar.android.auth.google.SignInProvider
-import com.leandrodev.contasdolar.get
-import com.leandrodev.contasdolar.getViewModel
 import com.leandrodev.contasdolar.android.model.User
+import com.leandrodev.shared.get
+import com.leandrodev.shared.getViewModel
 import com.leandrodev.ui.theme.HomeTheme
 
 @Composable
