+++ androidApp/src/main/java/com/leandrodev/contasdolar/android/auth/state/AuthStateContract.kt
@@ -2,7 +2,7 @@ package com.leandrodev.contasdolar.android.auth.state
 
 import com.leandrodev.contasdolar.MppViewModel
 import com.leandrodev.contasdolar.android.model.User
-import com.leandrodev.contasdolar.ViewState
+import com.leandrodev.shared.ViewState
 
 sealed interface AuthState {
     object Unknown : AuthState
