+++ androidApp/src/main/java/com/leandrodev/contasdolar/android/auth/state/AuthStateScreen.kt
@@ -6,8 +6,8 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import com.leandrodev.contasdolar.android.auth.signin.AuthScreen
-import com.leandrodev.contasdolar.getViewModel
 import com.leandrodev.contasdolar.android.home.HomeNavHost
+import com.leandrodev.shared.getViewModel
 
 @Composable
 fun AuthStateScreen(
