+++ shared/build.gradle.kts
@@ -2,6 +2,7 @@ plugins {
     id("com.leandrodev.mpplibrary")
 }
 
+group = "com.leandrodev.shared"
 version = "1.0"
 
 kotlin {
