--- shared/src/androidMain/kotlin/com/leandrodev/contasdolar/Platform.kt
@@ -1,5 +1,0 @@
-package com.leandrodev.contasdolar
-
-actual class Platform actual constructor() {
-    actual val platform: String = "Android ${android.os.Build.VERSION.SDK_INT}"
-}
