--- shared/src/commonMain/kotlin/com/leandrodev/contasdolar/Greeting.kt
@@ -1,7 +1,0 @@
-package com.leandrodev.contasdolar
-
-class Greeting {
-    fun greeting(): String {
-        return "Hello, ${Platform().platform}!"
-    }
-}
