--- shared/src/commonMain/kotlin/com/leandrodev/contasdolar/Platform.kt
@@ -1,5 +1,0 @@
-package com.leandrodev.contasdolar
-
-expect class Platform() {
-    val platform: String
-}
