--- shared/src/iosMain/kotlin/com/leandrodev/contasdolar/Platform.kt
@@ -1,7 +1,0 @@
-package com.leandrodev.contasdolar
-
-import platform.UIKit.UIDevice
-
-actual class Platform actual constructor() {
-    actual val platform: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
-}
