+++ shared/src/iosTest/kotlin/com/leandrodev/contasdolar/iosTest.kt
@@ -7,6 +7,6 @@ class IosGreetingTest {
 
     @Test
     fun testExample() {
-        assertTrue(Greeting().greeting().contains("iOS"), "Check iOS is mentioned")
+//        assertTrue(Greeting().greeting().contains("iOS"), "Check iOS is mentioned")
     }
 }
